refactor: move auth actions to app/auth, remove test pages
- Move actions/auth.ts to app/auth/actions.ts (co-locate with auth pages)
- Update imports in 6 auth page components
- Remove app/test-carousel/ and app/test/ dev pages
