+++ app/auth/signup/signup-content.tsx
@@ -8,7 +8,7 @@ import { Spinner } from "@/components/ui/spinner";
 import { FormHeading } from "@/components/ui/forms/FormHeading";
 import { Field, FieldGroup } from "@/components/ui/field";
 import { PasswordFields } from "@/components/auth/PasswordFields";
-import { signUpPublic } from "@/actions/auth";
+import { signUpPublic } from "@/app/auth/actions";
 
 type State = {
   message?: string;
